Initial upload
This is initial commit contains is a copy of the OER statistics
textbook titled: Introductory Statistics with Randomization and
Simulation, by Diaz, Barr, & Cetinkaya-Rundel.

The original source latex files were obtained from :
https://www.openintro.org/stat/textbook.php?stat_book=isrs

The original textbook is licensed under a Creative Commons license.

The original latex were converted to Markdown using pandoc. Some
minimal typesetting was completed on the first chapter.

The plan is to use this textbook as a starting point for developing
another OER textbook for tailored for undergraduate psychology students.

The Github changelog from this point forward reflects all of our
changes to the original textbook